+++ main.go
@@ -1,5 +1,20 @@
-package GriffinBackend
+package main
+
+import (
+	"fmt"
+	"log"
+	"os"
+
+	"github.com/joho/godotenv"
+)
+
+func init() {
+	err := godotenv.Load()
+	if err != nil {
+		log.Panicln(err)
+	}
+}
 
 func main() {
-
+	fmt.Println(os.Getenv("PORT"))
 }