+++ CHANGELOG.md
@@ -23,3 +23,26 @@ First Commit. Create Empty files + directory
 
 
 [Remove]
+
+
+# 0.0.2
+
+<p>
+
+Add JSON database structure. TODO: redis
+</p>
+
+[Add]
+- ./rdb
+  - dao_conn.go
+    - struct DataBase
+      - method Connect 
+  - dao_struct.go
+    - struct Employee
+    - struct EmployeePlural
+- ./main.go
+  - func init - load environment
+  
+[Change]
+- ./Dockerfile
+